Remove unused prefab and add packages to some scripts
Also cause SoundEffectPlayers to get volume from GameSettings on awake
