--- Assets/Prefabs/Managers/MetaManager.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f895580819da00242853bc2cc47a1c59
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
