+++ Assets/Scripts/Consequences/AutomaticFireConsequence.cs
@@ -1,4 +1,5 @@
 #nullable enable
+using Audio;
 using Triggers;
 using UnityEngine;
