+++ Assets/Scripts/Consequences/DeathConsequence.cs
@@ -1,4 +1,5 @@
 #nullable enable
+using Managers;
 using Triggers;
 using UnityEngine;
